`inE` robustness
Improve proof script robustness whereby selective rewriting depended on that `inE` expands `x \in A, when `A` is one of the collective `pred`combinators (e.g., `[predU _ & _ ]`), into an expression involving `pred_of_simpl (mem ..) x` subexpression, which need to be further simplified using `/=` or `inE`.
This will no longer be necessary with future versions of `ssrbool.v`, so this PR removes hacks that depend on this behavior, i.e., explicit `3!inE` repeat counts.