Adapt to Coq PR #18591: better refolding of addn induces "simpl never" now respected
An equality "2%N = (1 + \rank V)%N" was previously leading to "1%N = (\rank V)%N"
after injection but with the PR, it respects "simpl never" and only leads to
"1%N = (0 + \rank V)%N". So, we add a rewrite of addSn and add0n to
recover the previous behavior. This rewrite is additionally made
optional so as to preserve the compatibility.

Another simplification, in PFsection13.v, gives "1 + (c - 2)" which
now needs an extra add1n.